multiple backends: yet another change of plans
Try another aproach: define rf24_ops in each backend and export pointer to that
ops structure. If that backend is disabled then make that pointer NULL.

Do a lot of sanity checks in rf24_init to make sure that user is using valid mix
of library binary and its spi xfer ops.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>